+++ packages/core/src/RenderingEngine/BaseVolumeViewport.ts
@@ -31,6 +31,7 @@ import type {
   Point2,
   Point3,
   VOIRange,
+  ViewportProperties,
   VolumeViewportProperties,
 } from '../types';
 import { VoiModifiedEventDetail } from '../types/EventTypes';
@@ -45,6 +46,7 @@ import {
   invertRgbTransferFunction,
   triggerEvent,
   colormap as colormapUtils,
+  isImageActor,
 } from '../utilities';
 import { createVolumeActor } from './helpers';
 import volumeNewImageEventDispatcher, {
@@ -54,6 +56,7 @@ import Viewport from './Viewport';
 import type { vtkSlabCamera as vtkSlabCameraType } from './vtkClasses/vtkSlabCamera';
 import vtkSlabCamera from './vtkClasses/vtkSlabCamera';
 import transformWorldToIndex from '../utilities/transformWorldToIndex';
+import { getTransferFunctionNodes } from '../utilities/transferFunctionUtils';
 
 /**
  * Abstract base class for volume viewports. VolumeViewports are used to render
@@ -68,11 +71,16 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
   useCPURendering = false;
   use16BitTexture = false;
   private _FrameOfReferenceUID: string;
-  private inverted = false;
 
+  protected initialTransferFunctionNodes: any;
   // Viewport Properties
-  // TODO: similar to setVoi, this is only applicable to first volume
-  private VOILUTFunction: VOILUTFunctionType;
+  private globalDefaultProperties: VolumeViewportProperties;
+  private perVolumeIdDefaultProperties = new Map<
+    string,
+    VolumeViewportProperties
+  >();
+
+  private viewportProperties: VolumeViewportProperties = {};
 
   constructor(props: ViewportInput) {
     super(props);
@@ -217,8 +225,8 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
       voiLUTFunction = VOILUTFunctionType.LINEAR;
     }
     const { voiRange } = this.getProperties();
-    this.VOILUTFunction = voiLUTFunction;
     this.setVOI(voiRange, volumeId, suppressEvents);
+    this.viewportProperties.VOILUTFunction = voiLUTFunction;
   }
 
   /**
@@ -236,7 +244,6 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
     suppressEvents?: boolean
   ) {
     const applicableVolumeActorInfo = this._getApplicableVolumeActor(volumeId);
-
     if (!applicableVolumeActorInfo) {
       return;
     }
@@ -267,6 +274,8 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
     cfun.applyColorMap(colormapObj);
     cfun.setMappingRange(range[0], range[1]);
     volumeActor.getProperty().setRGBTransferFunction(0, cfun);
+
+    this.viewportProperties.colormap = colormap;
   }
 
   /**
@@ -283,6 +292,7 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
       return;
     }
     const { volumeActor } = applicableVolumeActorInfo;
+
     const ofun = vtkPiecewiseFunction.newInstance();
     if (typeof colormap.opacity === 'number') {
       const range = volumeActor
@@ -298,6 +308,8 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
       });
     }
     volumeActor.getProperty().setScalarOpacity(0, ofun);
+
+    this.viewportProperties.colormap.opacity = colormap.opacity;
   }
 
   /**
@@ -310,7 +322,7 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
    * @returns void
    */
   private setInvert(
-    invert: boolean,
+    inverted: boolean,
     volumeId?: string,
     suppressEvents?: boolean
   ) {
@@ -325,17 +337,18 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
     const cfun = this._getOrCreateColorTransferFunction(volumeIdToUse);
     invertRgbTransferFunction(cfun);
 
-    this.inverted = invert;
+    const { voiRange, VOILUTFunction, invert } =
+      this.getProperties(volumeIdToUse);
 
-    const { voiRange } = this.getProperties();
+    this.viewportProperties.invert = invert;
 
     if (!suppressEvents) {
       const eventDetail: VoiModifiedEventDetail = {
         viewportId: this.id,
         range: voiRange,
         volumeId: volumeIdToUse,
-        VOILUTFunction: this.VOILUTFunction,
-        invert: this.inverted,
+        VOILUTFunction: VOILUTFunction,
+        invert: invert,
         invertStateChanged: true,
       };
 
@@ -383,6 +396,7 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
 
     // @ts-ignore
     volumeProperty.setInterpolationType(interpolationType);
+    this.viewportProperties.interpolationType = interpolationType;
   }
 
   /**
@@ -415,9 +429,11 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
       voiRangeToUse = maxVoiRange;
     }
 
+    const { VOILUTFunction } = this.getProperties(volumeIdToUse);
+
     // scaling logic here
     // https://github.com/Kitware/vtk-js/blob/c6f2e12cddfe5c0386a73f0793eb6d9ab20d573e/Sources/Rendering/OpenGL/VolumeMapper/index.js#L957-L972
-    if (this.VOILUTFunction === VOILUTFunctionType.SAMPLED_SIGMOID) {
+    if (VOILUTFunction === VOILUTFunctionType.SAMPLED_SIGMOID) {
       const cfun = createSigmoidRGBTransferFunction(voiRangeToUse);
       volumeActor.getProperty().setRGBTransferFunction(0, cfun);
     } else {
@@ -433,6 +449,14 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
         .getProperty()
         .getRGBTransferFunction(0)
         .setRange(lower, upper);
+
+      if (!this.initialTransferFunctionNodes) {
+        const transferFunction = volumeActor
+          .getProperty()
+          .getRGBTransferFunction(0);
+        this.initialTransferFunctionNodes =
+          getTransferFunctionNodes(transferFunction);
+      }
     }
 
     if (!suppressEvents) {
@@ -440,15 +464,49 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
         viewportId: this.id,
         range: voiRange,
         volumeId: volumeIdToUse,
-        VOILUTFunction: this.VOILUTFunction,
+        VOILUTFunction: VOILUTFunction,
       };
 
       triggerEvent(this.element, Events.VOI_MODIFIED, eventDetail);
+    }
+
+    this.viewportProperties.voiRange = voiRangeToUse;
+  }
+
+  /**
+   * Update the default properties for the volume viewport on the volume
+   * @param ViewportProperties - The properties to set
+   * @param volumeId - The volume id to set the default properties for (if undefined, we set the global default viewport properties)
+   */
+  public setDefaultProperties(
+    ViewportProperties: VolumeViewportProperties,
+    volumeId?: string
+  ): void {
+    if (volumeId == null) {
+      this.globalDefaultProperties = ViewportProperties;
+    } else {
+      this.perVolumeIdDefaultProperties.set(volumeId, ViewportProperties);
+    }
+  }
+
+  /**
+   * Remove the global default properties of the viewport or remove default properties for a volumeId if specified
+   * @param volumeId If given, we remove the default properties only for this volumeId, if not
+   * the global default properties will be removed
+   */
+  public clearDefaultProperties(volumeId?: string): void {
+    if (volumeId == null) {
+      this.globalDefaultProperties = {};
+      this.resetProperties();
+    } else {
+      this.perVolumeIdDefaultProperties.delete(volumeId);
+      this.resetToDefaultProperties(volumeId);
     }
   }
 
   /**
    * Sets the properties for the volume viewport on the volume
+   * and if setProperties is called for the first time, the properties will also become the default one.
    * (if fusion, it sets it for the first volume in the fusion)
    *
    * @param VolumeViewportProperties - The properties to set
@@ -456,7 +514,7 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
    * @param [VolumeViewportProperties.VOILUTFunction] - Sets the voi mode (LINEAR, or SAMPLED_SIGMOID)
    * @param [VolumeViewportProperties.invert] - Inverts the color transfer function
    * @param [VolumeViewportProperties.colormap] - Sets the colormap
-   * @param [VolumeViewportProperties.preset] - Sets the colormap
+   * @param [VolumeViewportProperties.preset] - Sets the colormap preset
    * @param volumeId - The volume id to set the properties for (if undefined, the first volume)
    * @param suppressEvents - If true, the viewport will not emit events
    */
@@ -468,10 +526,23 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
       colormap,
       preset,
       interpolationType,
+      slabThickness,
     }: VolumeViewportProperties = {},
     volumeId?: string,
     suppressEvents = false
   ): void {
+    //If the viewport hasn't been initialized, we need to set the default properties
+    if (this.globalDefaultProperties == null) {
+      this.setDefaultProperties({
+        voiRange,
+        VOILUTFunction,
+        invert,
+        colormap,
+        preset,
+        slabThickness,
+      });
+    }
+
     // Note: colormap should always be done first, since we can then
     // modify the voiRange
 
@@ -494,13 +565,55 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
       this.setVOILUTFunction(VOILUTFunction, volumeId, suppressEvents);
     }
 
-    if (invert !== undefined && this.inverted !== invert) {
+    if (invert !== undefined) {
       this.setInvert(invert, volumeId, suppressEvents);
     }
 
     if (preset !== undefined) {
       this.setPreset(preset, volumeId, suppressEvents);
     }
+
+    if (slabThickness !== undefined) {
+      this.setSlabThickness(slabThickness);
+      //We need to set the current slab thickness here since setSlabThickness is define in VolumeViewport
+      this.viewportProperties.slabThickness = slabThickness;
+    }
+  }
+
+  /**
+   * Reset the viewport properties to the default values
+   */
+  public resetToDefaultProperties(volumeId: string): void {
+    const properties = this.globalDefaultProperties;
+
+    if (properties.colormap?.name) {
+      this.setColormap(properties.colormap, volumeId);
+    }
+    if (properties.colormap?.opacity != null) {
+      this.setOpacity(properties.colormap, volumeId);
+    }
+
+    if (properties.voiRange !== undefined) {
+      this.setVOI(properties.voiRange, volumeId);
+    }
+
+    if (properties.VOILUTFunction !== undefined) {
+      this.setVOILUTFunction(properties.VOILUTFunction, volumeId);
+    }
+
+    if (properties.invert !== undefined) {
+      this.setInvert(properties.invert, volumeId);
+    }
+
+    if (properties.slabThickness !== undefined) {
+      this.setSlabThickness(properties.slabThickness);
+      //We need to set the current slabthickness here since setSlabThickness is define in VolumeViewport
+      // this.currentViewportProperties.get(volumeId).slabThickness =
+      //   properties.slabThickness;
+      this.viewportProperties.slabThickness = properties.slabThickness;
+    }
+
+    this.render();
   }
 
   /**
@@ -530,13 +643,52 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
     }
 
     applyPreset(volumeActor, preset);
+
+    this.viewportProperties.preset = presetName;
   }
 
   /**
+   * Retrieve the viewport default properties
+   * @param volumeId If given, we retrieve the default properties of a volumeId if it exists
+   * If not given,we return the global properties of the viewport
+   * @returns default viewport properties including voi, invert, interpolation type, colormap
+   */
+  public getDefaultProperties = (
+    volumeId?: string
+  ): VolumeViewportProperties => {
+    let volumeProperties;
+    if (volumeId !== undefined) {
+      volumeProperties = this.perVolumeIdDefaultProperties.get(volumeId);
+    }
+
+    if (volumeProperties !== undefined) {
+      return volumeProperties;
+    }
+
+    return {
+      ...this.globalDefaultProperties,
+    };
+  };
+
+  /**
    * Retrieve the viewport properties
+   * @param volumeId - The volume id to get the properties for (if undefined, the first volume)
    * @returns viewport properties including voi, interpolation type: TODO: slabThickness, invert, rotation, flip
    */
-  public getProperties = (): VolumeViewportProperties => {
+  public getProperties = (volumeId?: string): VolumeViewportProperties => {
+    const applicableVolumeActorInfo = this._getApplicableVolumeActor(volumeId);
+    if (!applicableVolumeActorInfo) {
+      return;
+    }
+
+    const {
+      colormap,
+      VOILUTFunction,
+      interpolationType,
+      invert,
+      slabThickness,
+    } = this.viewportProperties;
+
     const voiRanges = this.getActors()
       .map((actorEntry) => {
         const volumeActor = actorEntry.actor as vtkVolume;
@@ -547,7 +699,7 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
         }
         const cfun = volumeActor.getProperty().getRGBTransferFunction(0);
         const [lower, upper] =
-          this.VOILUTFunction === 'SIGMOID'
+          this.viewportProperties?.VOILUTFunction === 'SIGMOID'
             ? getVoiFromSigmoidRGBTransferFunction(cfun)
             : cfun.getRange();
         return { volumeId, voiRange: { lower, upper } };
@@ -555,9 +707,15 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
       .filter(Boolean);
 
     const voiRange = voiRanges.length ? voiRanges[0].voiRange : null;
-    const VOILUTFunction = this.VOILUTFunction;
 
-    return { voiRange, VOILUTFunction, invert: this.inverted };
+    return {
+      colormap: colormap,
+      voiRange: voiRange,
+      VOILUTFunction: VOILUTFunction,
+      interpolationType: interpolationType,
+      invert: invert,
+      slabThickness: slabThickness,
+    };
   };
 
   /**
@@ -873,7 +1031,12 @@ abstract class BaseVolumeViewport extends Viewport implements IVolumeViewport {
    */
   private _setVolumeActors(volumeActorEntries: Array<ActorEntry>): void {
     // New volume actors implies resetting the inverted flag (i.e. like starting from scratch).
-    this.inverted = false;
+
+    for (let i = 0; i < volumeActorEntries.length; i++) {
+      const volumeId = volumeActorEntries[i].uid;
+
+      this.viewportProperties.invert = false;
+    }
     this.setActors(volumeActorEntries);
   }