+++ packages/core/src/RenderingEngine/StackViewport.ts
@@ -1,6 +1,7 @@
 import vtkDataArray from '@kitware/vtk.js/Common/Core/DataArray';
 import vtkImageData from '@kitware/vtk.js/Common/DataModel/ImageData';
 import type { vtkImageData as vtkImageDataType } from '@kitware/vtk.js/Common/DataModel/ImageData';
+import vtkColorMaps from '@kitware/vtk.js/Rendering/Core/ColorTransferFunction/ColorMaps';
 import _cloneDeep from 'lodash.clonedeep';
 import vtkCamera from '@kitware/vtk.js/Rendering/Core/Camera';
 import { vec2, vec3, mat4 } from 'gl-matrix';
@@ -19,6 +20,7 @@ import {
   imageIdToURI,
   isImageActor,
   actorIsA,
+  colormap as colormapUtils,
 } from '../utilities';
 import {
   Point2,
@@ -39,7 +41,7 @@ import {
   IStackViewport,
   VolumeActor,
   Mat3,
-  ColormapRegistration,
+  ColormapPublic,
   IImageCalibration,
 } from '../types';
 import { ViewportInput } from '../types/IViewport';
@@ -52,7 +54,6 @@ import {
   InterpolationType,
   RequestType,
   Events,
-  CalibrationTypes,
   VOILUTFunctionType,
 } from '../enums';
 import canvasToPixel from './helpers/cpuFallback/rendering/canvasToPixel';
@@ -79,6 +80,10 @@ import {
   ImagePlaneModule,
 } from '../types';
 import ViewportStatus from '../enums/ViewportStatus';
+import {
+  getTransferFunctionNodes,
+  setTransferFunctionNodes,
+} from '../utilities/transferFunctionUtils';
 
 const EPSILON = 1; // Slice Thickness
 
@@ -125,6 +130,13 @@ class StackViewport extends Viewport implements IStackViewport {
   private debouncedTimeout: number;
 
   // Viewport Properties
+  private globalDefaultProperties: StackViewportProperties;
+  private perImageIdDefaultProperties = new Map<
+    string,
+    StackViewportProperties
+  >();
+
+  private colormap: ColormapPublic | CPUFallbackColormapData;
   private voiRange: VOIRange;
   private voiUpdatedWithSetProperties = false;
   private VOILUTFunction: VOILUTFunctionType;
@@ -132,6 +144,7 @@ class StackViewport extends Viewport implements IStackViewport {
   private invert = false;
   // The initial invert of the image loaded as opposed to the invert status of the viewport itself (see above).
   private initialInvert = false;
+  private initialTransferFunctionNodes = null;
   private interpolationType: InterpolationType;
 
   // Helpers
@@ -251,14 +264,6 @@ class StackViewport extends Viewport implements IStackViewport {
   public getImageData: () => IImageData | CPUIImageData;
 
   /**
-   * Sets the colormap for the current viewport.
-   * @param colormap - The colormap data to use.
-   */
-  public setColormap: (
-    colormap: CPUFallbackColormapData | ColormapRegistration
-  ) => void;
-
-  /**
    * If the user has selected CPU rendering, return the CPU camera, otherwise
    * return the default camera
    * @returns The camera object.
@@ -364,6 +369,14 @@ class StackViewport extends Viewport implements IStackViewport {
   private setInterpolationType: (interpolationType: InterpolationType) => void;
 
   private setInvertColor: (invert: boolean) => void;
+
+  /**
+   * Sets the colormap for the current viewport.
+   * @param colormap - The colormap data to use.
+   */
+  private setColormap: (
+    colormap: CPUFallbackColormapData | ColormapPublic
+  ) => void;
 
   private initializeElementDisabledHandler() {
     eventTarget.addEventListener(
@@ -623,15 +636,60 @@ class StackViewport extends Viewport implements IStackViewport {
   }
 
   /**
-   * Sets the properties for the viewport on the default actor. Properties include
-   * setting the VOI, inverting the colors and setting the interpolation type, rotation
-   * @param voiRange - Sets the lower and upper voi
-   * @param invert - Inverts the colors
-   * @param interpolationType - Changes the interpolation type (1:linear, 0: nearest)
-   * @param rotation - image rotation in degrees
+   * Update the default properties of the viewport and add properties by imageId if specified
+   * @param ViewportProperties - The properties to set
+   * @param imageId If given, we set the default properties only for this image index, if not
+   * the default properties will be set for all imageIds
+   */
+  public setDefaultProperties(
+    ViewportProperties: StackViewportProperties,
+    imageId?: string
+  ): void {
+    if (imageId == null) {
+      this.globalDefaultProperties = ViewportProperties;
+    } else {
+      this.perImageIdDefaultProperties.set(imageId, ViewportProperties);
+
+      //If the viewport is the same imageIdI, we need to update the viewport
+      if (this.getCurrentImageId() === imageId) {
+        this.setProperties(ViewportProperties);
+      }
+    }
+  }
+
+  /**
+   * Remove the global default properties of the viewport or remove default properties for an imageId if specified
+   * @param imageId If given, we remove the default properties only for this imageID, if not
+   * the global default properties will be removed
+   */
+  public clearDefaultProperties(imageId?: string): void {
+    if (imageId == null) {
+      this.globalDefaultProperties = {};
+      this.resetProperties();
+    } else {
+      this.perImageIdDefaultProperties.delete(imageId);
+      this.resetToDefaultProperties();
+    }
+  }
+
+  /**
+   Configures the properties of the viewport.
+   This method allows customization of the viewport by setting attributes like
+   VOI (Value of Interest), color inversion, interpolation type, and image rotation.
+   If setProperties is called for the first time, the provided properties will
+   become the default settings for all images in the stack in case the resetPropertiese need to be called
+   @param properties - An object containing the properties to be set.
+   @param properties.colormap - Specifies the colormap for the viewport.
+   @param properties.voiRange - Defines the lower and upper Value of Interest (VOI) to be applied.
+   @param properties.VOILUTFunction - Function to handle the application of a lookup table (LUT) to the VOI.
+   @param properties.invert - A boolean value to toggle color inversion (true: inverted, false: not inverted).
+   @param properties.interpolationType - Determines the interpolation method to be used (1: linear, 0: nearest-neighbor).
+   @param properties.rotation - Specifies the image rotation angle in degrees.
+   @param suppressEvents - A boolean value to control event suppression. If true, the related events will not be triggered. Default is false.
    */
   public setProperties(
     {
+      colormap,
       voiRange,
       VOILUTFunction,
       invert,
@@ -643,6 +701,21 @@ class StackViewport extends Viewport implements IStackViewport {
     this.viewportStatus = this.csImage
       ? ViewportStatus.PRE_RENDER
       : ViewportStatus.LOADING;
+
+    if (this.globalDefaultProperties == null) {
+      this.setDefaultProperties({
+        colormap,
+        voiRange,
+        VOILUTFunction,
+        invert,
+        interpolationType,
+        rotation,
+      });
+    }
+
+    if (typeof colormap !== 'undefined') {
+      this.setColormap(colormap);
+    }
     // if voi is not applied for the first time, run the setVOI function
     // which will apply the default voi based on the range
     if (typeof voiRange !== 'undefined') {
@@ -671,11 +744,34 @@ class StackViewport extends Viewport implements IStackViewport {
   }
 
   /**
+   * Retrieve the viewport default properties
+   * @param imageId If given, we retrieve the default properties of an image index if it exists
+   * If not given,we return the global properties of the viewport
+   * @returns viewport properties including voi, invert, interpolation type, rotation, flip
+   */
+  public getDefaultProperties = (imageId?: string): StackViewportProperties => {
+    let imageProperties;
+    if (imageId !== undefined) {
+      imageProperties = this.perImageIdDefaultProperties.get(imageId);
+    }
+
+    if (imageProperties !== undefined) {
+      return imageProperties;
+    }
+
+    return {
+      ...this.globalDefaultProperties,
+      rotation: this.getRotation(),
+    };
+  };
+
+  /**
    * Retrieve the viewport properties
    * @returns viewport properties including voi, invert, interpolation type, rotation, flip
    */
   public getProperties = (): StackViewportProperties => {
     const {
+      colormap,
       voiRange,
       VOILUTFunction,
       interpolationType,
@@ -685,6 +781,7 @@ class StackViewport extends Viewport implements IStackViewport {
     const rotation = this.getRotation();
 
     return {
+      colormap,
       voiRange,
       VOILUTFunction,
       interpolationType,
@@ -733,6 +830,52 @@ class StackViewport extends Viewport implements IStackViewport {
     }
     this.setInterpolationType(InterpolationType.LINEAR);
     this.setInvertColor(this.initialInvert);
+
+    const transferFunction = this.getTransferFunction();
+    setTransferFunctionNodes(
+      transferFunction,
+      this.initialTransferFunctionNodes
+    );
+  }
+
+  public resetToDefaultProperties(): void {
+    this.cpuRenderingInvalidated = true;
+    this.viewportStatus = ViewportStatus.PRE_RENDER;
+
+    this.fillWithBackgroundColor();
+
+    if (this.useCPURendering) {
+      this._cpuFallbackEnabledElement.renderingTools = {};
+    }
+
+    const currentImageId = this.getCurrentImageId();
+    const properties =
+      this.perImageIdDefaultProperties.get(currentImageId) ||
+      this.globalDefaultProperties;
+
+    if (properties.colormap?.name) {
+      this.setColormap(properties.colormap);
+    }
+
+    let voiRange;
+    if (properties.voiRange == undefined) {
+      // if not set via setProperties; if it is not a PT image or is not prescaled,
+      // use the voiRange for the current image from its metadata if found
+      // otherwise, use the cached voiRange
+      voiRange = this._getVOIRangeForCurrentImage();
+    } else {
+      voiRange = properties.voiRange;
+    }
+
+    this.setVOI(voiRange);
+
+    if (this.getRotation() !== 0) {
+      this.setRotation(0);
+    }
+    this.setInterpolationType(InterpolationType.LINEAR);
+    this.setInvertColor(false);
+
+    this.render();
   }
 
   private _setPropertiesFromCache(): void {
@@ -1017,7 +1160,7 @@ class StackViewport extends Viewport implements IStackViewport {
       ? vec3.negate(vec3.create(), this.initialViewUp)
       : this.initialViewUp;
 
-    this.setCamera({
+    this.setCameraNoEvent({
       viewUp: initialViewUp as Point3,
     });
 
@@ -1040,6 +1183,7 @@ class StackViewport extends Viewport implements IStackViewport {
 
     // @ts-ignore
     volumeProperty.setInterpolationType(interpolationType);
+
     this.interpolationType = interpolationType;
   }
 
@@ -1091,6 +1235,7 @@ class StackViewport extends Viewport implements IStackViewport {
       if ((!this.invert && invert) || (this.invert && !invert)) {
         invertRgbTransferFunction(tfunc);
       }
+
       this.invert = invert;
     }
   }
@@ -1143,6 +1288,21 @@ class StackViewport extends Viewport implements IStackViewport {
     if (!suppressEvents) {
       triggerEvent(this.element, Events.VOI_MODIFIED, eventDetail);
     }
+  }
+
+  private getTransferFunction() {
+    const defaultActor = this.getDefaultActor();
+
+    if (!defaultActor) {
+      return;
+    }
+
+    if (!isImageActor(defaultActor)) {
+      return;
+    }
+    const imageActor = defaultActor.actor as ImageActor;
+
+    return imageActor.getProperty().getRGBTransferFunction(0);
   }
 
   private setVOIGPU(voiRange: VOIRange, options: SetVOIOptions = {}): void {
@@ -1204,6 +1364,8 @@ class StackViewport extends Viewport implements IStackViewport {
       }
 
       imageActor.getProperty().setRGBTransferFunction(0, transferFunction);
+      this.initialTransferFunctionNodes =
+        getTransferFunctionNodes(transferFunction);
     }
 
     if (!isSigmoidTFun) {
@@ -2089,7 +2251,7 @@ class StackViewport extends Viewport implements IStackViewport {
 
   private _getInitialVOIRange(image: IImage) {
     if (this.voiRange && this.voiUpdatedWithSetProperties) {
-      return this.voiRange;
+      return this.globalDefaultProperties.voiRange;
     }
     const { windowCenter, windowWidth } = image;
 
@@ -2169,6 +2331,17 @@ class StackViewport extends Viewport implements IStackViewport {
       imageIdIndex
     );
 
+    //Check if there is any existing specific options for images if not we don't
+    //want to re-render the viewport to its default properties
+    if (this.perImageIdDefaultProperties.size >= 1) {
+      const defaultProperties = this.perImageIdDefaultProperties.get(imageId);
+      if (defaultProperties !== undefined) {
+        this.setProperties(defaultProperties);
+      } else if (this.globalDefaultProperties !== undefined) {
+        this.setProperties(this.globalDefaultProperties);
+      }
+    }
+
     return imageId;
   }
 
@@ -2244,7 +2417,7 @@ class StackViewport extends Viewport implements IStackViewport {
 
     const targetImageId = imageIds[newTargetImageIdIndex];
 
-    const imageAlreadyLoaded = cache.isImageIdCached(targetImageId);
+    const imageAlreadyLoaded = cache.isLoaded(targetImageId);
 
     // If image is already cached we want to scroll right away; however, if it is
     // not cached, we can debounce the scroll event to avoid firing multiple scroll
@@ -2647,6 +2820,7 @@ class StackViewport extends Viewport implements IStackViewport {
   }
 
   private setColormapCPU(colormapData: CPUFallbackColormapData) {
+    this.colormap = colormapData;
     const colormap = getColormap(colormapData.name, colormapData);
 
     this._cpuFallbackEnabledElement.viewport.colormap = colormap;
@@ -2658,23 +2832,31 @@ class StackViewport extends Viewport implements IStackViewport {
     this.render();
   }
 
-  private setColormapGPU(colormap: ColormapRegistration) {
+  private setColormapGPU(colormap: ColormapPublic) {
     const ActorEntry = this.getDefaultActor();
     const actor = ActorEntry.actor as ImageActor;
     const actorProp = actor.getProperty();
     const rgbTransferFunction = actorProp.getRGBTransferFunction();
 
+    const colormapObj =
+      colormapUtils.getColormap(colormap.name) ||
+      vtkColorMaps.getPresetByName(colormap.name);
+
     if (!rgbTransferFunction) {
       const cfun = vtkColorTransferFunction.newInstance();
-      const voiRange = this._getVOIRangeForCurrentImage();
-      cfun.applyColorMap(colormap);
-      cfun.setMappingRange(voiRange.lower, voiRange.upper);
+      cfun.applyColorMap(colormapObj);
+      cfun.setMappingRange(this.voiRange.lower, this.voiRange.upper);
       actorProp.setRGBTransferFunction(0, cfun);
     } else {
-      rgbTransferFunction.applyColorMap(colormap);
+      rgbTransferFunction.applyColorMap(colormapObj);
+      rgbTransferFunction.setMappingRange(
+        this.voiRange.lower,
+        this.voiRange.upper
+      );
       actorProp.setRGBTransferFunction(0, rgbTransferFunction);
     }
 
+    this.colormap = colormap;
     this.render();
   }