+++ packages/core/src/RenderingEngine/VolumeViewport.ts
@@ -1,9 +1,11 @@
 import vtkPlane from '@kitware/vtk.js/Common/DataModel/Plane';
+import vtkVolume from '@kitware/vtk.js/Rendering/Core/Volume';
+
 import { vec3 } from 'gl-matrix';
 
 import cache from '../cache';
 import { MPR_CAMERA_VALUES, RENDERING_DEFAULTS } from '../constants';
-import { BlendModes, OrientationAxis } from '../enums';
+import { BlendModes, OrientationAxis, Events } from '../enums';
 import type {
   ActorEntry,
   IImageVolume,
@@ -12,10 +14,16 @@ import type {
   Point3,
 } from '../types';
 import type { ViewportInput } from '../types/IViewport';
-import { actorIsA, getClosestImageId } from '../utilities';
+import {
+  actorIsA,
+  getClosestImageId,
+  isImageActor,
+  triggerEvent,
+} from '../utilities';
 import BaseVolumeViewport from './BaseVolumeViewport';
 import setDefaultVolumeVOI from './helpers/setDefaultVolumeVOI';
-import vtkVolume from '@kitware/vtk.js/Rendering/Core/Volume';
+import { setTransferFunctionNodes } from '../utilities/transferFunctionUtils';
+import { ImageActor } from '../types/IActor';
 
 /**
  * An object representing a VolumeViewport. VolumeViewports are used to render
@@ -271,6 +279,11 @@ class VolumeViewport extends BaseVolumeViewport {
    * the slab thickness to (if not provided, all actors will be affected).
    */
   public setSlabThickness(slabThickness: number, filterActorUIDs = []): void {
+    if (slabThickness < 0.1) {
+      // Cannot render zero thickness
+      slabThickness = 0.1;
+    }
+
     let actorEntries = this.getActors();
 
     if (filterActorUIDs && filterActorUIDs.length > 0) {
@@ -383,6 +396,31 @@ class VolumeViewport extends BaseVolumeViewport {
       );
     }
     setDefaultVolumeVOI(volumeActor.actor as vtkVolume, imageVolume, false);
+
+    if (isImageActor(volumeActor)) {
+      setTransferFunctionNodes(
+        (volumeActor.actor as ImageActor)
+          .getProperty()
+          .getRGBTransferFunction(0),
+        this.initialTransferFunctionNodes
+      );
+    }
+
+    const range = (volumeActor.actor as vtkVolume)
+      .getProperty()
+      .getRGBTransferFunction(0)
+      .getMappingRange();
+
+    const eventDetails = {
+      viewportId: volumeActor.uid,
+      range: {
+        lower: range[0],
+        upper: range[1],
+      },
+      volumeId: volumeActor.uid,
+    };
+
+    triggerEvent(this.element, Events.VOI_MODIFIED, eventDetails);
   }
 }