+++ packages/tools/src/utilities/getModalityUnit.ts
@@ -28,8 +28,14 @@ function _handlePTModality(imageId: string, options: ModalityUnitOptions) {
     return 'SUV';
   }
 
-  const petSeriesModule = metaData.get('petSeriesModule', imageId);
-  return petSeriesModule?.units || 'unitless';
+  const generalSeriesModule = metaData.get('generalSeriesModule', imageId);
+
+  // it might be possible that the referenceImageId is not the one
+  // that is being displayed. So we need to get the modality from imageId again
+  if (generalSeriesModule?.modality === 'PT') {
+    const petSeriesModule = metaData.get('petSeriesModule', imageId);
+    return petSeriesModule?.units || 'unitless';
+  }
 }
 
 export { getModalityUnit, ModalityUnitOptions };